requirements.txt
PySide6
platformdirs
astropy
auto-stretch
rclone-python
pillow
opencv-python
